Migrate libultra/os message queue functions to rsbs/ (WP-009)
Move message queue reference implementations from games/oot/ and games/mm/
to unified rsbs/src/libultra/os/:
- createmesgqueue.c
- sendmesg.c
- recvmesg.c
- jammesg.c
- seteventmesg.c

These files are commented out in rsbs/CMakeLists.txt as they are reference
implementations - runtime functionality is provided by libultraship.
