+++ rsbs/src/libultra/os/createmesgqueue.c
@@ -1,4 +1,13 @@
-#include "global.h"
+/**
+ * @file createmesgqueue.c
+ * @brief OS message queue creation - unified from OoT/MM decomp
+ *
+ * Original sources:
+ *   - games/oot/src/libultra/os/createmesgqueue.c
+ *   - games/mm/src/libultra/os/createmesgqueue.c
+ */
+
+#include "ultra64.h"
 
 void osCreateMesgQueue(OSMesgQueue* mq, OSMesg* msg, s32 count) {
     mq->mtqueue = (OSThread*)__osThreadTail;
